requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
google-auth==2.13.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
google-api-python-client==2.88.0
python-dotenv==0.21.1  # For loading environment variables from a .env file
